+++ src/main/java/io/sentry/cli/SentryCliRunner.java
@@ -42,8 +42,7 @@ public class SentryCliRunner {
   public @Nullable String runSentryCli(
       final @NotNull String sentryCliCommand, final boolean failOnError)
       throws MojoExecutionException {
-    final boolean isWindows =
-        System.getProperty("os.name").toLowerCase(Locale.ROOT).startsWith("windows");
+    final boolean isWindows = isWindows();
 
     final @NotNull String executable = isWindows ? "cmd.exe" : "/bin/sh";
     final @NotNull String cArg = isWindows ? "/c" : "-c";
@@ -65,16 +64,17 @@ public class SentryCliRunner {
                       attributes(
                           attribute("executable", executable),
                           attribute("failOnError", String.valueOf(failOnError)),
-                          attribute("output", logFile.getAbsolutePath())),
+                          attribute("output", escape(logFile.getAbsolutePath()))),
                       element(name("arg"), attributes(attribute("value", cArg))),
                       element(
                           name("arg"),
                           attributes(
                               attribute(
                                   "value",
-                                  getCliPath(mavenProject, sentryCliExecutablePath)
-                                      + " "
-                                      + sentryCliCommand)))))),
+                                  wrapForWindows(
+                                      escape(getCliPath(mavenProject, sentryCliExecutablePath))
+                                          + " "
+                                          + sentryCliCommand))))))),
           executionEnvironment(mavenProject, mavenSession, pluginManager));
 
       return collectAndMaybePrintOutput(logFile, debugSentryCli);
@@ -89,6 +89,38 @@ public class SentryCliRunner {
       throw e;
     } catch (IOException e) {
       throw new RuntimeException(e);
+    }
+  }
+
+  private static boolean isWindows() {
+    final boolean isWindows =
+        System.getProperty("os.name").toLowerCase(Locale.ROOT).startsWith("windows");
+    return isWindows;
+  }
+
+  private @Nullable String wrapForWindows(final @Nullable String toWrap) {
+    // Wrap whole command in double quotes as Windows cmd will remove the first and last double
+    // quote
+    if (toWrap != null && isWindows()) {
+      return "\"" + toWrap + "\"";
+    } else {
+      return toWrap;
+    }
+  }
+
+  public @Nullable String escape(final @Nullable String escapePath) {
+    if (escapePath == null) {
+      return null;
+    }
+    if (isWindows()) {
+      // Wrap paths that contain a whitespace in double quotes
+      // For some reason wrapping paths that do not contain a whitespace leads to an error
+      if (escapePath.contains(" ")) {
+        return "\"" + escapePath + "\"";
+      }
+      return escapePath;
+    } else {
+      return escapePath.replaceAll(" ", "\\\\ ");
     }
   }